+++ packages/atproto/test/src/sessions/session_service_test.dart
@@ -109,4 +109,65 @@ void main() {
       );
     });
   });
+
+  group('.refreshSession', () {
+    test('normal case', () async {
+      final sessions = SessionsService(
+        did: 'test',
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/sessions/data/refresh_session.json',
+        ),
+      );
+
+      final response = await sessions.refreshSession(
+        refreshJwt: '',
+      );
+
+      expect(response, isA<core.XRPCResponse>());
+      expect(response.data, isA<Session>());
+    });
+
+    test('when unauthorized', () async {
+      final sessions = SessionsService(
+        did: 'test',
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await sessions.refreshSession(refreshJwt: ''),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final sessions = SessionsService(
+        did: 'test',
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await sessions.refreshSession(refreshJwt: ''),
+      );
+    });
+  });
 }