+++ packages/atproto_core/lib/src/client/client.dart
@@ -20,7 +20,8 @@ abstract class Client {
   Future<xrpc.XRPCResponse<T>> post<T>(
     final xrpc.NSID methodId, {
     required final String service,
-    required final dynamic body,
+    final Map<String, String>? headers,
+    final dynamic body,
     final xrpc.To<T>? to,
     required final Duration timeout,
     final xrpc.PostClient? postClient,